[Sehajpreet] Add: Find the discounted price paid by the students

+++ DiscountCalculator.java
@@ -1,16 +1,17 @@
+
 public class DiscountCalculator {
     public static void main(String[] args) {
-        
-        int fee = 125000;
-        int discountPercent = 10;
+        //the original fee and discount percentage as float
+        float fee = 125000f;
+        float discountPercent = 10f;
 
         // Calculate the discount amount
-        double discount = fee * discountPercent / 100.0;
+        float discount = fee * discountPercent / 100.0f;
 
         // Calculate the final fee after discount
-        double finalFee = fee - discount;
+        float finalFee = fee - discount;
 
-        
+       
         System.out.println("The discount amount is INR " + discount + 
                            " and final discounted fee is INR " + finalFee);
     }
